TDB-1474 Fix for Random lookup support in MappedReadbackStrategy
In order to cut down on the heap overhead of the TreeMap/MarkerEntry overhead
in the MappedReadbackStrategy, I implemented a specialized long/long ordered 
array. The array is broken into chunks, and within each chunk, the longs are stored as
deltas from their respective first values in the chunk.

The array is expected to be used for ever-increasing keys and values, without key 
dups, where the longs within a chunk are no more than Integer.MAX_VALUE from 
the base. This is ideal for a per segment file lsn/chunk index. 
 
 The long/long array was used to implement a MarkerDictionary, used by the 
 MappedReadbackStrategy to provide random access to Chunks by LSN.

svn path=/fast-restartable-store/trunk/; revision=49367