TDB-2710 AbstractChunk.findEnd was overly expensive
Basically, because it always looped from the beginning of the
buffer list. Since it is called in a loop, this lead to repeated
N squared operations, which is less than stellar.

By tracking where the traversal left off, you can then restart
it there, skipping all of this.

Also, split findEnd into two separate methods, one for put and
one for get, as it made the logic more clear.

svn path=/fast-restartable-store/trunk/; revision=52967