Add latency tracking for FSync in FRS.
Here we add latency tracking. We track it in the fastest way
I could come up with, a log base 2 set of bins. This allows
for very fast recording of the duration, but it does limit what the
bins can be since they must be powers of 2.

By default, the latency bins are reported every 5 minutes; they are reset
on report.

The "com.terracottatech.frs.ReportFsyncLatenciesInSecs" property
controls the period with which latencies are reported. <=0 will
disable reporting, other wise the number is interpreted as the
number of seconds between reports. Default is 300.


svn path=/fast-restartable-store/trunk/; revision=54469